+++ crates/rtlola-hir/src/modes/dependencies.rs
@@ -89,7 +89,7 @@ pub(crate) type Streamdependencies = HashMap<SRef, Vec<(SRef, Vec<(Origin, Strea
 /// Represents all transitive dependencies between streams
 pub(crate) type Transitivedependencies = HashMap<SRef, Vec<SRef>>;
 /// Represents all dependencies between streams in which a window lookup is used
-pub(crate) type Windowdependencies = HashMap<SRef, Vec<(SRef, WRef)>>;
+pub(crate) type Windowdependencies = HashMap<SRef, Vec<(SRef, Origin, WRef)>>;
 
 pub(crate) trait ExtendedDepGraph {
     /// Returns a new [dependency graph](DependencyGraph), in which all edges representing a negative offset lookup are deleted
@@ -230,23 +230,23 @@ impl DepAnaTrait for DepAna {
             .map_or(Vec::new(), |accesses| accesses.to_vec())
     }
 
-    fn aggregated_by(&self, who: SRef) -> Vec<(SRef, WRef)> {
+    fn aggregated_by(&self, who: SRef) -> Vec<(SRef, Origin, WRef)> {
         self.aggregated_by
             .get(&who)
             .map_or(Vec::new(), |aggregated_by| {
                 aggregated_by
                     .iter()
-                    .map(|(sref, wref)| (*sref, *wref))
-                    .collect::<Vec<(SRef, WRef)>>()
+                    .map(|(sref, origin, wref)| (*sref, *origin, *wref))
+                    .collect::<Vec<_>>()
             })
     }
 
-    fn aggregates(&self, who: SRef) -> Vec<(SRef, WRef)> {
+    fn aggregates(&self, who: SRef) -> Vec<(SRef, Origin, WRef)> {
         self.aggregates.get(&who).map_or(Vec::new(), |aggregates| {
             aggregates
                 .iter()
-                .map(|(sref, wref)| (*sref, *wref))
-                .collect::<Vec<(SRef, WRef)>>()
+                .map(|(sref, origin, wref)| (*sref, *origin, *wref))
+                .collect::<Vec<_>>()
         })
     }
 
@@ -395,9 +395,9 @@ impl DepAna {
             spec.all_streams().map(|sr| (sr, Vec::new())).collect();
         let mut direct_accessed_by: HashMap<SRef, Vec<(SRef, Origin, StreamAccessKind)>> =
             spec.all_streams().map(|sr| (sr, Vec::new())).collect();
-        let mut aggregates: HashMap<SRef, Vec<(SRef, WRef)>> =
+        let mut aggregates: HashMap<SRef, Vec<(SRef, Origin, WRef)>> =
             spec.all_streams().map(|sr| (sr, Vec::new())).collect();
-        let mut aggregated_by: HashMap<SRef, Vec<(SRef, WRef)>> =
+        let mut aggregated_by: HashMap<SRef, Vec<(SRef, Origin, WRef)>> =
             spec.all_streams().map(|sr| (sr, Vec::new())).collect();
         edges.iter().for_each(|(src, w, tar)| {
             let cur_accesses = direct_accesses.get_mut(src).unwrap();
@@ -412,12 +412,12 @@ impl DepAna {
             }
             if let Some(wref) = w.window() {
                 let cur_aggregates = aggregates.get_mut(src).unwrap();
-                if !cur_aggregates.contains(&(*tar, wref)) {
-                    cur_aggregates.push((*tar, wref));
+                if !cur_aggregates.contains(&(*tar, w.origin, wref)) {
+                    cur_aggregates.push((*tar, w.origin, wref));
                 }
                 let cur_aggregates_by = aggregated_by.get_mut(tar).unwrap();
-                if !cur_aggregates_by.contains(&(*src, wref)) {
-                    cur_aggregates_by.push((*src, wref));
+                if !cur_aggregates_by.contains(&(*src, w.origin, wref)) {
+                    cur_aggregates_by.push((*src, w.origin, wref));
                 }
             }
         });
@@ -710,9 +710,9 @@ mod tests {
             deps.aggregates.iter().for_each(|(sr, aggregates_hir)| {
                 let aggregates_reference = aggregates.get(sr).unwrap();
                 assert_eq!(aggregates_hir.len(), aggregates_reference.len(), "test");
-                aggregates_hir
-                    .iter()
-                    .for_each(|lookup| assert!(aggregates_reference.contains(lookup)));
+                aggregates_hir.iter().for_each(|(sr, _, wref)| {
+                    assert!(aggregates_reference.contains(&(*sr, *wref)))
+                });
             });
             deps.aggregated_by
                 .iter()
@@ -723,9 +723,9 @@ mod tests {
                         aggregated_by_reference.len(),
                         "test"
                     );
-                    aggregated_by_hir
-                        .iter()
-                        .for_each(|lookup| assert!(aggregated_by_reference.contains(lookup)));
+                    aggregated_by_hir.iter().for_each(|(sr, _, wref)| {
+                        assert!(aggregated_by_reference.contains(&(*sr, *wref)))
+                    });
                 });
         } else {
             assert!(dependencies.is_none())